파일업로드 준비

+++ src/main/webapp/WEB-INF/templates/board/index.jsp
@@ -10,6 +10,13 @@
 <c:url var="deleteUrl" value="/board/delete" />
 <c:url var="searchUrl" value="/board/list" />
 <script>
+    window.addEventListener("DOMContentLoaded", function() {
+        const keywordEl = document.getElementById("keyword");
+
+        keywordEl.addEventListener("change", function(e) {
+            alert("키워드를 바꾸었으므로 페이지는 1로 초기화");
+        });
+    });
     function fn_search(page) {
 
         let keyword = document.getElementById("keyword").value;
@@ -72,14 +79,13 @@
                 </td>
                 <td>
                     <a href="${updateUrl}/${board.artNo}"> <fmt:message key="수정하기" /></a> |
-                    <a href="${deleteUrl}/${board.artNo}"> <fmt:message key="삭제하기" /></a>
+                    <a href="${deleteUrl}/${board.artNo}" onclick="return confirm('게시글을 정말 삭제하시겠습니까')"> <fmt:message key="삭제하기" /></a>
                 </td>
             </tr>
             </c:forEach>
             </tbody>
         </table>
         <div class="paging">
-            <b>${currentPage}</b>
             <c:forEach var="i" begin="1" end="${maxPage}">
                 <c:if test="${i!=currentPage}" >
                     <a href="#" onclick="fn_search(${i})">${i}</a>
